FIX for car and treetrunk collision + added every sprite + fixed bird movement
Car and treetrunk collision now work when passing the second time.

Added every sprite on the map, ready for background image.

Bird movement updated to move further + move upscreen at a certain point.

TODO: ADD Death scene, ADD Coin sound, ADD Death sound, ADD Timer as score and function to deduct coins off of the timer, ADD Finish scene that shows the time, ADD Background

+++ crossy_road/src/sprites/bird.h
@@ -49,9 +49,9 @@ public:
     void tick(u16 keys);
 
     ///Declaration start position bird
-    double xPosition = (GBA_SCREEN_WIDTH/2 - 16);
-    double yPosition = (GBA_SCREEN_HEIGHT - 32);
-    double virtualYPos = yPosition;
+    int xPosition = (GBA_SCREEN_WIDTH/2 - 16);
+    int yPosition = (GBA_SCREEN_HEIGHT - 32);
+    int virtualYPos = yPosition;
 };
